requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
protobuf
grpcio-tools
playwright
pyperclip